requirements.txt
anyio==3.7.0
APScheduler==3.6.3
asgiref==3.7.2
async-timeout==4.0.2
attrs==23.1.0
autobahn==23.6.2
Automat==22.10.0
beautifulsoup4==4.12.2
cachetools==4.2.2
certifi==2023.7.22
cffi==1.15.1
channels==4.0.0
channels-redis==4.1.0
charset-normalizer==3.1.0
constantly==15.1.0
cryptography==41.0.5
daphne==4.0.0
Django==4.2.2
django-auth-ldap==4.6.0
django-bootstrap4==23.2
djangorestframework==3.14.0
exceptiongroup==1.1.1
h11==0.14.0
httpcore==0.17.2
httpx==0.24.1
hyperlink==21.0.0
idna==3.4
incremental==22.10.0
ldap3==2.9.1
msgpack==1.0.5
numpy==1.25.1
Pillow==10.0.0
psycopg==3.1.12
psycopg-binary==3.1.12
psycopg2-binary==2.9.9
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
pyOpenSSL==23.2.0
python-ldap==3.4.3
python-telegram-bot==13.7
pytz==2023.3
redis==4.6.0
requests==2.31.0
service-identity==23.1.0
six==1.16.0
sniffio==1.3.0
soupsieve==2.4.1
sqlparse==0.4.4
tornado==6.3.2
Twisted==22.10.0
txaio==23.1.1
typing_extensions==4.6.3
tzlocal==5.0.1
urllib3==2.0.7
zope.interface==6.0
